feat: msg.text() upd: change bang dir
