+++ plugins/corpus/bang.json
@@ -16,7 +16,8 @@
       "cv": "相羽爱奈",
       "motto": ["做好为Roselia赌上一切的觉悟了吗？","目指すのは頂点、ただひとつよ","愿我们能够永远作为Roselia走下去","你也，向往着音乐吧"]
     },
-    {"name": ["sayo","纱夜","冰川纱夜","氷川 紗夜","忧伤节拍器","Hikawa Sayo","HikawaSayo","薯条妹",
+    {
+      "name": ["sayo","纱夜","冰川纱夜","氷川 紗夜","忧伤节拍器","Hikawa Sayo","HikawaSayo","薯条妹",
         "琴行老板","冰川打火机","342g"],
       "birthday": "03-20",
       "cv": "工藤晴香",
@@ -24,7 +25,8 @@
     }
   ],
   "Pastel*Palettes": [
-    {"name": ["hina","日菜","冰川日菜","氷川 日菜","Hikawa Hina","HikawaHina","噜","噜噜噜","隔壁的小天才"],
+    {
+      "name": ["hina","日菜","冰川日菜","氷川 日菜","Hikawa Hina","HikawaHina","噜","噜噜噜","隔壁的小天才"],
       "birthday": "03-20",
       "cv": "小泽亚李",
       "motto": ["噜~♪","るんっ♪","噜噜噜~♪","噜♪起来了！","お姉ちゃん~~","我觉得开始噜♪了！"]
